Give map, filter and sink the ability to use sub-flows
`map`, `filter` and `sink` used to receive as input the whole object
flowing through their pipeline. This change allows all three to use
the `args` keyword argument to specify a single key or a tuple of keys
which will be used to extract items from the object in the pipeline to
be used as arguments to the component.

Additionally, `map` has an optional `out` keyword argument which
specifies where (under what item key) the values returned by the
transforming function should be placed inside the object flowing
through the pipe. Once again, this can be a single key, when the
transformer function's output is a single object, or a tuple of keys
if the transformer function returns multiple results.

These features make it easy to split and join multiple synchronized
streams. This is done by sending all the streams down a sigle pipe in
some mapping object such as a dictionary, with each stream identified
by some key. The `args` parameter can be used to specify which
sub-streams should be consumed by the component (`map`, `filter` or
`sink`) while the `out` parameter can be used to specify into which
sub-streams the results should be placed.

By picking multiple substreams with `args` substreams can be joined.

By using a previously non-existent key in `out` a new substream is
added to the flow, which is equivalent to fork.

Additionaly the `item` keyword argument to `map` allows applying the
transformation to a specific substream. For example,

    map(op, args='abc', out='abc')

is equivalent to

    map(op, item='abc)